fix(trader): convert timestamp parameters to milliseconds in tests
- Cast serverTime mock response to int64 in Binance futures test mocks
- Convert FilledAt timestamp to milliseconds using time.UnixMilli() in binance_sync_e2e_test
- Update GetLastFillTimeByExchange return value handling to work with millisecond timestamps
- Convert CreatedAt field to UnixMilli() for fake fill creation in test data
- Add UnixMilli() conversion for all trade creation timestamps in exchange_sync_test
- Convert timestamp parameters to milliseconds in hyperliquid_sync_test trade processing
- Fix comment alignment in TestTrade struct definition
- Ensure consistent millisecond timestamp handling across all exchange sync tests

+++ trader/binance_futures_test.go
@@ -251,7 +251,7 @@ func NewBinanceFuturesTestSuite(t *testing.T) *BinanceFuturesTestSuite {
 		// Mock ServerTime - /fapi/v1/time
 		case path == "/fapi/v1/time":
 			respBody = map[string]interface{}{
-				"serverTime": 1234567890000,
+				"serverTime": int64(1234567890000),
 			}
 
 		// Default: empty response
@@ -326,7 +326,7 @@ func TestNewFuturesTrader(t *testing.T) {
 		switch path {
 		case "/fapi/v1/time":
 			respBody = map[string]interface{}{
-				"serverTime": 1234567890000,
+				"serverTime": int64(1234567890000),
 			}
 		case "/fapi/v1/positionSide/dual":
 			respBody = map[string]interface{}{
